+++ lib/heirloom/cli/tag.rb
@@ -17,8 +17,7 @@ module Heirloom
         ensure_valid_options :provided => @opts,
                              :required => [:name, :id, :attribute, :value],
                              :config   => @config
-        ensure_valid_region :region => @opts[:metadata_region],
-                            :config => @config
+        ensure_valid_metadata_region @config
         ensure_domain_exists :name => @opts[:name], :config => @config
 
         @archive = Archive.new :name   => @opts[:name],
@@ -27,7 +26,7 @@ module Heirloom
         ensure_archive_exists :archive => @archive,
                               :config => @config
       end
-      
+
       def tag
         unless @archive.exists?
           @logger.error "Archive does not exist"
@@ -65,7 +64,7 @@ EOS
           opt :aws_secret_key, "AWS Secret Access Key", :type  => :string, 
                                                         :short => :none
           opt :use_iam_profile, "Use IAM EC2 Profile", :short => :none
-          opt :environment, "Environment (defined in ~/.heirloom.yml)", :type => :string
+          opt :environment, "Environment (defined in heirloom config file)", :type => :string
         end
       end
     end