+++ lib/heirloom/cli/teardown.rb
@@ -14,7 +14,8 @@ module Heirloom
         @config = load_config :logger => @logger,
                               :opts   => @opts
 
-        ensure_valid_options :provided => @opts, 
+        ensure_valid_metadata_region @config
+        ensure_valid_options :provided => @opts,
                              :required => [:name],
                              :config   => @config
 
@@ -82,7 +83,7 @@ EOS
           opt :aws_secret_key, "AWS Secret Access Key", :type => :string, 
                                                         :short => :none
           opt :use_iam_profile, "Use IAM EC2 Profile", :short => :none
-          opt :environment, "Environment (defined in ~/.heirloom.yml)", :type => :string
+          opt :environment, "Environment (defined in heirloom config file)", :type => :string
         end
       end